fix: address code review findings on Venmo vault flow
- Thread venmoPendingStorage through the full component tree so the
  redirect-restoration test exercises firstUpdated instead of calling
  the private method via `as any`
- Extract MockStorage to test/helpers/mock-storage.ts and remove the
  duplicate definition from both test files
- Add comment in checkVenmoRestoration explaining why startPayment() is
  called (not instance.tokenize()) after hasTokenizationResult()
- Improve Google Pay suppression comment to name the upstream property
  path so a future breakage is traceable
- Convert catch (e: any) to catch (e: unknown) with type narrowing in
  startVenmoPayment and checkVenmoRestoration
